requirements.txt
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-phonenumber-field==7.0.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.4
